requirements.txt
pandas
matplotlib
requests
google-generativeai
python-dotenv
xlsxwriter